add reset in (Field)ODEEventDetector
