+++ hipparchus-ode/src/changes/changes.xml
@@ -50,6 +50,9 @@ If the output is not quite correct, check for invisible trailing spaces!
   </properties>
   <body>
     <release version="4.0" date="TBD" description="TBD">
+      <action dev="serrof" type="add" issue="issues/382">
+        Added reset mechanism in ODE event detection.
+      </action>
       <action dev="serrof" type="update" issue="issues/372">
         Rename RungeKutta(Field)Integrator as FixedStepRungeKutta(Field)Integrator.
       </action>
