+++ hipparchus-ode/src/test/java/org/hipparchus/ode/events/FieldDetectorBasedEventStateTest.java
@@ -80,7 +80,7 @@ class FieldDetectorBasedEventStateTest {
 
         @Override
         public FieldODEEventHandler<Complex> getHandler() {
-            return null;
+            return (s, e, d) -> Action.CONTINUE;
         }
 
         @Override
